Use InputFile for mpqInput

+++ Models/Progress.cs
@@ -4,9 +4,9 @@ public record struct Progress
 {
     public string Message { get; init; }
 
-    public int Total { get; init; }
+    public long Total { get; init; }
 
-    public int BytesLoaded { get; init; }
+    public long BytesLoaded { get; init; }
 
     public double PercentLoaded => Total != 0 ? Round(100d * BytesLoaded / Total) : 100d;
 }
